move `JsonbVal` to types mod
