+++ src/common/src/test_utils/rand_array.rs
@@ -25,9 +25,10 @@ use rand::rngs::SmallRng;
 use rand::{Rng, SeedableRng};
 
 use crate::array::serial_array::Serial;
-use crate::array::{Array, ArrayBuilder, ArrayRef, JsonbVal, ListValue, StructValue};
-use crate::types::num256::Int256;
-use crate::types::{Date, Decimal, Interval, NativeType, Scalar, Time, Timestamp};
+use crate::array::{Array, ArrayBuilder, ArrayRef, ListValue, StructValue};
+use crate::types::{
+    Date, Decimal, Int256, Interval, JsonbVal, NativeType, Scalar, Time, Timestamp,
+};
 
 pub trait RandValue {
     fn rand_value<R: Rng>(rand: &mut R) -> Self;
