+++ src/expr/benches/expr.rs
@@ -24,10 +24,8 @@ use std::cell::RefCell;
 use criterion::async_executor::FuturesExecutor;
 use criterion::{criterion_group, criterion_main, Criterion};
 use risingwave_common::array::*;
-use risingwave_common::types::test_utils::IntervalTestExt;
-use risingwave_common::types::{
-    DataType, DataTypeName, Date, Decimal, Interval, Time, Timestamp, F32, F64,
-};
+use risingwave_common::types::interval::test_utils::IntervalTestExt;
+use risingwave_common::types::*;
 use risingwave_expr::agg::{build as build_agg, AggArgs, AggCall};
 use risingwave_expr::expr::*;
 use risingwave_expr::sig::agg::agg_func_sigs;
