+++ src/expr/src/vector_op/arithmetic_op.rs
@@ -409,10 +409,10 @@ pub fn sqrt_decimal(expr: Decimal) -> Result<Decimal> {
 mod tests {
     use std::str::FromStr;
 
-    use num_traits::Float;
-    use risingwave_common::types::num256::{Int256, Int256Ref};
-    use risingwave_common::types::test_utils::IntervalTestExt;
-    use risingwave_common::types::{Date, Decimal, Interval, Scalar, Timestamp, F32, F64};
+    use risingwave_common::types::interval::test_utils::IntervalTestExt;
+    use risingwave_common::types::{
+        Date, Decimal, Int256, Int256Ref, Interval, Scalar, Timestamp, F32, F64,
+    };
 
     use super::*;
